+++ components/about-me/AboutMe.vue
@@ -1,9 +1,7 @@
 <template>
   <SectionWrapper id="about-me" icon="me" :title="$t('aboutMe')">
-    <div class="flex self-center justify-center w-full h-full">
-      <AboutMeTerminalFrame :buttons-clickable="true">
-        <AboutMeTerminalContent />
-      </AboutMeTerminalFrame>
-    </div>
+    <AboutMeTerminalFrame :buttons-clickable="true">
+      <AboutMeTerminalContent />
+    </AboutMeTerminalFrame>
   </SectionWrapper>
 </template>